+++ assets/block-library/grd/grd-item/grd-item-save.js
@@ -16,9 +16,9 @@ import { getStyle, getStyleFromObject } from '../grd-edit';
 
 /**
  * Generate block HTML to save to the database
- * @param {Object} root0
- * @param {Object} root0.attributes
- * @param {string} root0.className
+ * @param {Object} props
+ * @param {Object} props.attributes
+ * @param {string} props.className
  */
 const GridItemSave = ({ attributes, className }) => {
 	const { span, col, row, spanRow, vAlign, variation } = attributes;