+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 7.2.1 - October 29 2024
+* Standardize background color functionality.
+* Removed unused comments.php (use component-library/blog/comments-view.php).
+* Renamed date block hook function to prevent generic function name clashes.
+* Tweaked alignment for more control over block gap values.
+* Updated CSS custom properties and several theme.json values to unify some common values.
+
 ## 7.2.0 - October 10 2024
 * Prevent PostCSS from allowing nested CSS ([see this issue](https://github.com/Yoast/wordpress-seo/issues/20535)).
 * Update alignment styling (e.g. content, wide, full) so back-end and front-end match.